# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iterra API",
    "description": "AI Content Strategy Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/content/calendar": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Calendar",
        "operationId": "calendar_api_v1_content_calendar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEventResponse"
                  },
                  "title": "Response Calendar Api V1 Content Calendar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/posts": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Posts",
        "operationId": "posts_api_v1_analytics_posts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "linkedin",
              "title": "Platform"
            }
          },
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostWithAnalysis"
                  },
                  "title": "Response Posts Api V1 Analytics Posts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/analyze/{post_id}": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_v1_analytics_analyze__post_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          },
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calendar/generate": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Generate content calendar",
        "description": "Produces a structured weekly-style plan. By default this is a deterministic mock plan suitable for demos and CI. Set `USE_ITERRA_AI_CALENDAR=true` with `ANTHROPIC_API_KEY` configured to use `iterra_ai.CalendarEngine` (falls back to mock on LLM failure).",
        "operationId": "generate_calendar_api_v1_calendar_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendar/": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List Calendars",
        "operationId": "list_calendars_api_v1_calendar__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarOutput"
                  },
                  "title": "Response List Calendars Api V1 Calendar  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "platform",
          "status",
          "starts_at",
          "content"
        ],
        "title": "CalendarEventResponse"
      },
      "CalendarInput": {
        "properties": {
          "niche": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Niche"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Platforms"
          },
          "posting_frequency": {
            "type": "integer",
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Posting Frequency"
          },
          "historical_posts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Historical Posts"
          }
        },
        "type": "object",
        "required": [
          "niche",
          "posting_frequency"
        ],
        "title": "CalendarInput"
      },
      "CalendarOutput": {
        "properties": {
          "content_plan": {
            "items": {
              "$ref": "#/components/schemas/ContentSlot"
            },
            "type": "array",
            "title": "Content Plan"
          }
        },
        "type": "object",
        "required": [
          "content_plan"
        ],
        "title": "CalendarOutput"
      },
...
      "ContentSlot": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "date",
          "platform",
          "content_type",
          "topic",
          "goal"
        ],
        "title": "ContentSlot"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostAnalysisResponse": {
        "properties": {
          "post_id": {
            "type": "string",
            "title": "Post Id"
          },
          "hook_score": {
            "type": "integer",
            "title": "Hook Score"
          },
          "tone_match_score": {
            "type": "integer",
            "title": "Tone Match Score"
          },
          "structure_score": {
            "type": "integer",
            "title": "Structure Score"
          },
          "cta_effectiveness": {
            "type": "string",
            "title": "Cta Effectiveness"
          },
          "top_strength": {
            "type": "string",
            "title": "Top Strength"
          },
          "top_improvement": {
            "type": "string",
            "title": "Top Improvement"
          },
          "predicted_engagement": {
            "type": "string",
            "title": "Predicted Engagement"
          },
          "rewrite_suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewrite Suggestion"
          }
        },
        "type": "object",
        "required": [
          "post_id",
          "hook_score",
          "tone_match_score",
          "structure_score",
          "cta_effectiveness",
          "top_strength",
          "top_improvement",
          "predicted_engagement"
        ],
        "title": "PostAnalysisResponse"
      },
      "PostWithAnalysis": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "likes": {
            "type": "integer",
            "title": "Likes"
          },
          "comments": {
            "type": "integer",
            "title": "Comments"
          },
          "shares": {
            "type": "integer",
            "title": "Shares"
          },
          "engagement_rate": {
            "type": "number",
            "title": "Engagement Rate"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "platform",
          "content",
          "likes",
          "comments",
          "shares",
          "engagement_rate"
        ],
        "title": "PostWithAnalysis"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}